+++ rabbitDb-impl/src/test/java/com/farmerworking/db/rabbitDb/impl/utils/ErrorRandomAccessFile.java
@@ -6,7 +6,7 @@ import org.apache.commons.lang3.tuple.Pair;
 
 public class ErrorRandomAccessFile implements RandomAccessFile {
     @Override
-    public Pair<Status, String> read(long offset, long n) {
+    public Pair<Status, String> read(int offset, int n) {
         return Pair.of(Status.iOError("read error"), null);
     }
 }
